Add hint module
Compiler hints are often useful in order to improve the performance of
the compiled binary. The Argent Core library provides a Compiler Hints
Module to take advantage of some of the most useful hints. The following
hints have been made available through this module:
  - `arc_hot`
  - `arc_cold`
  - `arc_pure`
  - `arc_likely()`
  - `arc_unlikely()`

The interface of this module has been documented, and examples of the
usage of this interface have been included in the example/hint.h file.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>