Add types module
An important objective of the Argent Core Library is to be as portable
as possible, and to be backward compatible with the C89 standard. With
this in mind, the types module has been added to provide a consistent
and portable means of referring to the primitive data types.

The types module has been added with the following interface:
  - `agc_bool` type
  - `agc_word` type
  - `agc_word_8` type
  - `agc_word_16` type
  - `agc_word_32` type
  - `agc_word_64` type
  - `agc_size` type
  - `agc_index` type
  - `agc_int` type
  - `agc_int_8` type
  - `agc_int_16` type
  - `agc_int_32` type
  - `agc_int_64` type
  - `agc_uint` type
  - `agc_uint_8` type
  - `agc_uint_16` type
  - `agc_uint_32` type
  - `agc_uint_64` type
  - `agc_float` type
  - `agc_float_32` type
  - `agc_float_64` type

The namespace of the Library has also been changed from `arc_` to `agc_`
to better reflect the name `argent`.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>