Simplify Argent namespace
Experience has shown that it makes sense to use a common namespace for
the Argent Libraries rather than specific ones for each. Accordingly,
the `agc_` namespace used earlier has been renamed to `ag_`. The
namespace in the other Argent Libraries need to be modified accordingly.

As a result of this simplification, it is now possible to define the
`ag_string` type within the Types Module of the Core Library.

The examples have also been updated accordingly, and a new example has
been added for `ag_string`.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>